+++ src/apis/Login/auth.ts
@@ -1,26 +1,21 @@
 import { axiosInstance } from '../axios.ts';
 import type { loginResponse } from '../../types/LoginPage/auth.ts';
 
-export const loginWithGoogle = async (authCode: string, state?: string): Promise<loginResponse> => {
-  // GET /api/auth/login/google?code={authorization_code} 요청
-  const response = await axiosInstance.get<loginResponse>(`/api/auth/login/google`, {
-    params: {
-      code: authCode,
-      state: state,
-    },
-  });
-  console.log('구글 로그인 응답:', response.data);
-  return response.data;
+export const postGoogleAuthCode = async (authCode: string) => {
+  const { data } = await axiosInstance.post<loginResponse>('/api/auth/google/token', { code: authCode });
+  console.log('Google login response:', data);
+  return data;
 };
 
-export const loginWithNaver = async (authCode: string, state?: string): Promise<loginResponse> => {
-  // GET /api/auth/login/naver?code={authorization_code} 요청
-  const response = await axiosInstance.get<loginResponse>(`/api/auth/login/naver`, {
-    params: {
-      code: authCode,
-      state: state,
-    },
-  });
+export const postNaverAuthCode = async (authCode: string) => {
+  const { data } = await axiosInstance.post<loginResponse>('/api/auth/naver/token', { code: authCode });
+  console.log('Naver login response:', data);
+  return data;
+};
 
-  return response.data;
+//카카오는 request body에 인가 코드뿐만 아니라 추가적인 정보까지 담을 듯 합니다
+export const postKakaoAuthCode = async (authCode: string) => {
+  const { data } = await axiosInstance.post<loginResponse>('/api/auth/kakao/token', { code: authCode });
+  console.log('Kakao login response:', data);
+  return data;
 };
