+++ src/pages/callback/GoogleCallbackPage.tsx
@@ -1,47 +1,71 @@
-import React, { useEffect } from 'react';
-import { useSearchParams, useNavigate } from 'react-router-dom';
-import { useAuth } from '@/context/AuthContext'; // 기존에 만들어둔 AuthContext
 
-const GoogleCallbackPage = () => {
-  // URL의 쿼리 파라미터(?code=...)를 쉽게 다루게 해주는 훅
-  const [searchParams] = useSearchParams();
-  // 페이지 이동을 위한 훅
-  const navigate = useNavigate();
-  // 기존에 만들어둔 login 함수를 가져오기 위한 훅
+import { useEffect } from 'react';
+import { useLocation } from 'react-router-dom';
+import { useAuth } from '@/context/AuthContext';
+
+
+const SocialLoginCallback = () => {
+  const location = useLocation();
   const { login } = useAuth();
 
-  // 이 컴포넌트가 화면에 렌더링되면 단 한 번만 실행됩니다.
   useEffect(() => {
-    const processLogin = async () => {
-      // 1. URL에서 'code' 라는 이름의 파라미터 값을 추출합니다.
-      const authCode = searchParams.get('code');
+    const handleSocialLogin = async () => {
+      // 네이버 처리 로직
+      // 네이버는 JS SDK 사용 시, 콜백 URL의 해시(#) 부분에 토큰을 담아 반환합니다.
+      // 예: http://.../callback#access_token=NAVER_TOKEN...
+      if (location.hash) {
+        // location.hash는 '#'을 포함하므로, substring(1)으로 제거합니다.
+        const hash = location.hash.substring(1);
+        const params = new URLSearchParams(hash);
+        const naverToken = params.get('access_token');
 
-      // 2. 만약 authCode가 존재한다면, 로그인 절차를 진행합니다.
-      if (authCode) {
-        try {
-          // 3. 기존에 구현해둔 login 함수를 호출합니다!
-          await login('google', authCode);
-
-          // 4. 로그인 성공 시, 메인 페이지로 이동시킵니다.
-          navigate('/');
-        } catch (error) {
-          // 백엔드에서 에러가 발생한 경우
-          console.error('로그인 처리 중 에러 발생:', error);
-          alert('로그인에 실패했습니다. 다시 시도해주세요.');
-          navigate('/login'); // 에러 발생 시 로그인 페이지로 이동
+        if (naverToken) {
+          console.log('✅ 네이버 토큰(인가 코드 역할) 받기 성공:', naverToken);
+          await login('naver', naverToken);
+          return 'success'; // 처리 성공을 나타내는 문자열 반환
         }
-      } else {
-        // 'code' 파라미터 없이 비정상적으로 접근한 경우
-        alert('잘못된 접근입니다.');
-        navigate('/login');
       }
+
+      // 카카오 처리 로직
+      // 카카오는 표준 OAuth 2.0 방식에 따라, 콜백 URL의 쿼리 스트링(?)에 인가 코드를 담아 반환합니다.
+      // 예: http://.../callback?code=KAKAO_CODE...
+      if (location.search) {
+        const params = new URLSearchParams(location.search);
+        const kakaoCode = params.get('code');
+
+        if (kakaoCode) {
+          console.log('✅ 카카오 인가 코드 받기 성공:', kakaoCode);
+          await login('kakao', kakaoCode);
+          return 'success'; // 처리 성공을 나타내는 문자열 반환
+        }
+      }
+
+      // 위 두 경우에 모두 해당하지 않으면 실패로 간주합니다.
+      return 'fail';
     };
 
-    processLogin();
-  }, []); // 빈 배열을 전달하여 최초 렌더링 시에만 실행되도록 설정
+    handleSocialLogin()
+      .then((result) => {
+        if (result === 'success') {
+          // 로그인 성공 시, 팝업을 띄운 부모 창이 있다면 해당 창을 메인으로 이동시키고
+          // 현재 팝업창은 닫습니다.
+          if (window.opener) {
+            window.opener.location.href = '/';
+          }
+          window.close();
+        } else {
+          // 성공적으로 처리되지 않았을 경우 에러를 발생시킵니다.
+          throw new Error('유효한 소셜 로그인 정보를 찾을 수 없습니다.');
+        }
+      })
+      .catch((error) => {
+        console.error('❗️ 소셜 로그인 처리 과정에서 에러 발생', error);
+        alert('로그인에 실패했습니다. 다시 시도해주세요.');
+        window.close();
+      });
+  }, [location, login]); // location과 login 함수가 변경될 때만 이 효과를 다시 실행합니다.
 
-  // 사용자에게는 이 메시지만 보입니다.
-  return <div>구글 로그인 처리 중입니다. 잠시만 기다려주세요...</div>;
+  return <div>소셜 로그인 처리 중입니다...</div>;
 };
 
-export default GoogleCallbackPage;
+export default SocialLoginCallback;
